pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "attack-path-sim"
version = "0.0.1"
description = "Lightweight attack path simulation + Monte Carlo risk estimation for prototype security scenarios"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name="Jo Y."}]
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.5.0",
]

[project.scripts]
attack-path-sim = "attack_path_sim.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
